requirements.txt
flask==3.0.2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.3.2
python-dotenv==1.0.1
requests==2.31.0
bottleneck==1.3.6
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0 